+++ backend/corpora/guardianobserver/guardianobserver.py
@@ -15,7 +15,7 @@ from ianalyzer_readers.xml_tag import Tag
 
 from django.conf import settings
 
-from es.es_update import update_document
+from indexing.run_update_task import update_document
 from addcorpus.python_corpora import extract
 from addcorpus.python_corpora import filters
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition, until, after, string_contains, consolidate_start_end_years
